Added all Kolkata work
This is all code from the group Kolkata in course 02267 software
development of web services copied into new Git repo for further
changes. The reason for not making the changes in the original git
repository is that the course has finished and I wanted to further
improve the logic for personal development reasons.